Now with functions to do time evolution. Hard coded in project.

+++ FCIForBE.cpp
@@ -590,8 +590,8 @@ std::vector< double > BEImpurityFCI(Eigen::MatrixXd &DensityMatrix, InputObj &In
 	/* This section is for direct diagonalization. Uncomment if desired. */
 	Timer = omp_get_wtime();
 	std::cout << "FCI: Beginning Direct Diagonalization... ";
-	Eigen::MatrixXf HamDense = Ham;
-	Eigen::SelfAdjointEigenSolver< Eigen::MatrixXf > HamEV;
+	Eigen::MatrixXd HamDense = Ham;
+	Eigen::SelfAdjointEigenSolver< Eigen::MatrixXd > HamEV;
 	HamEV.compute(HamDense);
 
 	std::cout << " done" << std::endl;
